+++ Package.swift
@@ -12,7 +12,7 @@ let package = Package(
     platforms: [
         .iOS(.v18),
         .macOS(.v15),
-        .visionOS(.v2)
+        .visionOS(.v1)
     ],
     products: [
         .library(